requirements.txt
pyperclip>=1.8.2
pystray>=0.19.4
Pillow>=10.0.0
pyobjc-framework-Cocoa>=9.0; sys_platform == "darwin"